Streamline interactor forms.
This makes a small but important change to the semantics of
with-current-interactor: it must evaluate (current-interactor)
after an exception occurs, and not before evaluating its thunk.

I've also removed with-interactor, since this form doesn't play
nice with the parameter-aware form.

All in all, this gets us back to SRFI 249's interactor interface,
approximately.