Generalize interactors.
Interactors now take a (usually composite) restarter condition,
instead of a list of (simple) restarters. This allows the user of
an interactor to inspect the condition raised by the triggering
exception. Information such as the original irritants and message
may be very important in diagnosing a problem, so we do not want
to strip it off a composite restarter before passing it to an
interactor.

Thanks to Arthur A. Gleckler for pointing out the lack of this
information in the interactor examples.